feat: add clusters + refactor existing code + docs
