+++ Readme.md
@@ -1,9 +1,14 @@
 # Simple CRUD API without frameworks
 implemented base architecture, users endpoint and test for it.
 
+## How to start
+- node version must be 16 LTS
+- `npm intall`
+
 ## Commands
 - `npm run start:dev` - start dev server
 - `npm run start:dev:inspect` - start dev server in debug mode
 - `npm run start:prod` - start production build
-- `npm run lint` - check and show all eslint issue
+- `npm run lint` - check and show all eslint issues
 - `npm run test:e2e` - run e2e tests(api)
+
