+++ index.ts
@@ -1,0 +1,3 @@
+import { boot } from './src/main';
+
+boot({ scale: true });
