+++ nodemon.json
@@ -6,5 +6,5 @@
 	"ignore": [
 		"src/**/*.spec.ts"
 	],
-	"exec": "ts-node --files ./src/main.ts"
+	"exec": "ts-node --files ./index.ts"
 }
