+++ package.json
@@ -2,13 +2,13 @@
   "name": "crud-api",
   "version": "0.0.1",
   "description": "Simple Crud API without frameworks",
-  "main": "src/main.js",
+  "main": "index.js",
 	"type": "commonjs",
   "scripts": {
     "build": "tsc",
     "start:dev": "nodemon",
-    "start:dev:inspect": "nodemon -e ts,json --exec node --inspect=localhost:6667 -r ts-node/register src/main.ts",
-    "start:prod": "npm run build && node ./dist/src/main.js",
+    "start:dev:inspect": "nodemon -e ts,json --exec node --inspect=localhost:6667 -r ts-node/register index.ts",
+    "start:prod": "npm run build && node ./dist/index.js",
     "lint": "eslint ./src/**",
     "test:e2e": "jest --config jest.e2e.config.ts"
   },
