+++ src/models/User.ts
@@ -1,5 +1,16 @@
-import { UserDB } from '../db/User';
+import { InMemoryDatabase } from '../core/InMemoryDatabase';
 
-export default class UserModel extends UserDB {
-	// logic related to user, must be placed here
+export default class UserModel extends InMemoryDatabase {
+	methods: InMemoryDatabase;
+	requiredParams: string[];
+
+	constructor() {
+		super();
+		this.requiredParams = ['username', 'age', 'hobbies'];
+		this.methods = this.workWith('users');
+	}
+
+	clearData(): void {
+		this.clearEntitiesData('users');
+	}
 }
